scripts/requirements.txt
ruamel.yaml>=0.17.21
python-dotenv>=1.0.0
httpx>=0.28.0
tenacity>=9.0.0
pydantic>=2.10.0
